+++ Main/Include/fluid.h
@@ -20,7 +20,7 @@ class fluid : public entity
   virtual void Be();
   void Save(outputfile&) const;
   void Load(inputfile&);
-  void Draw(bitmap*, vector2d, ushort, bool, bool) const;
+  void Draw(bitmap*, vector2d, ulong, bool, bool) const;
   bitmap* GetPicture() const { return Picture; }
   material* GetMaterial() const { return Material; }
   virtual square* GetSquareUnder() const { return SquareUnder; }
